libertine: initial open-source code base for LeEco phones
This commit is a squash of the initial tree commits written the
previous weeks. The build system is far from complete and the main
addons are still buggy, FreedomOS was used as example and for the main
components in AROMA. Their scripts were a little bit blurry to
incorporate "as-is" on a simplified base which uses different components
like EUI/Oxygen.

The main differences happen at Google Apps that we don't support on EUI
and at the scripts.

As of now:

- AROMA installer refactored for the better.
- Simplified scripts for FreedomOS.
- Built from outside as per files in configuration/*.ini.
- Different modes: build and analisis.
- In 'build' mode the ROM gets downloaded, extracted, modified and
repacked with AROMA.
- In 'analisis' mode the ROM gets downloaded, extracted and system
mounted for analisis at analyze/system/*.
- Google Velvet and Phonesky from upstream OpenGapps, latest GmsCore
from USA base release.

Check TODOs at configuration/*.ini.

Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>